--- src/pages/advisors.jsx
@@ -1,11 +1,0 @@
-import React, { useEffect } from "react";
-
-const paperURL = "/team#advisors";
-
-export default function Advisors() {
-  useEffect(() => {
-    window.location.replace(paperURL);
-  });
-
-  return <p>Redirecting...</p>;
-}
