+++ src/pages/team.jsx
@@ -1,40 +1,12 @@
-import React from "react";
-import { graphql } from "gatsby";
-import SEO from "../components/SEO";
-import Layout from "../components/Layout";
-import ValuesSection from "../components/team/ValuesSection";
-import TeamSection from "../components/team/TeamSection";
-import Supporters from "../components/team/Supporters";
+import React, { useEffect } from "react";
+import { useIntl } from "gatsby-plugin-intl";
+const paperURL = "/ourOrg#our-team";
 
-export default function Team({ data }) {
-  const images = {};
+export default function Paper() {
+  const intl = useIntl();
+  useEffect(() => {
+    window.location.replace(`/${intl.locale}${paperURL}`);
+  });
 
-  // add the images to an object to use in gatsby image
-  data.allFile.edges.forEach((edge) => (images[edge.node.name] = edge.node));
-
-  return (
-    <Layout>
-      <SEO title="Team | Virufy" />
-      <ValuesSection images={images} />
-      <TeamSection images={images} />
-      <Supporters />
-    </Layout>
-  );
+  return <p>Redirecting...</p>;
 }
-
-export const query = graphql`
-  {
-    allFile(filter: { relativeDirectory: { eq: "images/team" } }) {
-      edges {
-        node {
-          name
-          childImageSharp {
-            fluid {
-              ...GatsbyImageSharpFluid
-            }
-          }
-        }
-      }
-    }
-  }
-`;
